[fix] 채팅 인원 수 제한 test 01

+++ src/main/java/com/project/raidho/domain/chat/ChatDto/RoomDetailResponseDto.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class RoomDetailResponseDto {
     private Long roomMasterId;
     private String roomName;
-    private Long people;
+    private int people;
 }
